server/requirements.txt
blinker==1.4
click==8.0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
py-bcrypt==0.4
SQLAlchemy==1.4.31
Werkzeug==2.0.3
